[skeleton] Use haversine distance in greedy policy
