+++ source/state.py
@@ -1,34 +1,6 @@
-import numpy as np
 from itertools import product
 from source.scenario import Scenario
-
-
-
-def haversine_distance_deprecated(lat_1, lng_1, lat_2, lng_2):
-    size = len(lat_1) * len(lat_2)
-    # ToDo: Implement function
-    return np.ones(shape=(1, size)) * 100
-
-
-def haversine_distance(lat1, lon1, lat2, lon2):
-    """
-    Calculate the great circle distance between two points
-    on the earth (specified in decimal degrees)
-
-    All args must be of equal length.
-
-    """
-    lon1, lat1, lon2, lat2 = map(np.radians, [lon1, lat1, lon2, lat2])
-
-    dlon = lon2 - lon1
-    dlat = lat2 - lat1
-
-    a = np.sin(dlat/2.0)**2 + np.cos(lat1) * np.cos(lat2) * np.sin(dlon/2.0)**2
-
-    c = 2 * np.arcsin(np.sqrt(a))
-    km = 6367 * c
-    return km
-
+from source.utils import haversine_distance
 
 class State:
     def __init__(self, scenario: Scenario):
@@ -50,8 +22,8 @@ class State:
         nearest_order = dict()
         total_cost = 0
         for j, courier_id in enumerate(actions):
-            nearest_order[courier_id] = {'order_id': None, 'distance': 99999, 'lat': None,
-                                         'lng': None}  # ToDO: write infinity
+            nearest_order[courier_id] = {'order_id': None, 'distance': float('inf'), 'lat': None,
+                                         'lng': None}
             for i, order in enumerate(self.orders):
                 d = distance_array[(i + j * len(self.orders))]
                 if d < nearest_order[courier_id]['distance']:
